move FilterConstants into constants folder
